backend/fastapi_service/requirements.txt
fastapi==0.120.2
uvicorn==0.38.0
pydantic==2.12.3
openai==2.6.1
httpx==0.27.2
python-dotenv==1.0.0
redis==4.5.5
fastapi-limiter==0.1.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1